separated CC backend into .h and .cc files (breaks linking for now because of duplicate synbols)
