+++ openql/CMakeLists.txt
@@ -31,8 +31,8 @@ SET_PROPERTY(SOURCE openql.i PROPERTY SWIG_FLAGS -castmode)
 SWIG_ADD_LIBRARY(openql LANGUAGE python SOURCES openql.i TYPE SHARED)
 
 if (${CMAKE_SYSTEM_NAME} MATCHES "Darwin")
-    SWIG_LINK_LIBRARIES(openql ${LEMON_LIBRARIES})
+    SWIG_LINK_LIBRARIES(openql ${LEMON_LIBRARIES} ql)
     set_target_properties(_openql PROPERTIES LINK_FLAGS "-undefined dynamic_lookup")
 else ()
-    SWIG_LINK_LIBRARIES(openql ${PYTHON_LIBRARIES} ${LEMON_LIBRARIES})
+    SWIG_LINK_LIBRARIES(openql ${PYTHON_LIBRARIES} ${LEMON_LIBRARIES} ql)
 endif()
