+++ ql/utils.h
@@ -16,6 +16,8 @@
 #include <iterator>
 #include <string>
 #include <sstream>
+#include <fstream>
+#include <iostream>
 #include <utility>
 #include <vector>
